+++ examples/demo-rollup/src/mock_rollup.rs
@@ -97,8 +97,7 @@ impl FullNodeBlueprint<Native> for MockDemoRollup<Native> {
             sequencer_db,
             da_service,
             da_sync_state,
-            &rollup_config.sequencer,
-            &rollup_config.runner,
+            rollup_config,
         )
         .await?;